% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunkwise.R
\name{chunkwise}
\alias{chunkwise}
\title{Text file processing in chunks}
\usage{
chunkwise(file_name, chunksize, todo)
}
\arguments{
\item{file_name}{String of path and name of the text file to be processed in chunks.}

\item{chunksize}{Integer of the maximal number of rows to be processed in each chunk.}

\item{todo}{A function describing the process to be applied to a chunk.}
}
\value{
A data.table including the results from the chunks.
}
\description{
Given a large text file (possibly larger than available RAM), this function applies
a user-defined function on the data. The processing is done in a selected size of smaller chunks,
so the user does not have to read the entire data set before manipulating it.The function supports
baseR, dplyr and data.table syntax as well.
}
\examples{
\dontrun{
# subset with data.table syntax
filter1 <- function(x){x[COLUMN_1 == 'Some Value', .(COLUMN_1, COLUMN_2)]}

dt1 <- chunkwise(file_name = 'C:/Users/USERNAME/Desktop/largedata.csv',
                 chunksize = 10000L,
                 todo = filter1)

# subset with dplyr syntax
filter2 <- function(x){x \%>\%
                         filter(COLUMN_1 == 'Some Value') \%>\%
                         select(COLUMN_1, COLUMN_2)
                         }

dt2 <- chunkwise(file_name = 'C:/Users/USERNAME/Desktop/largedata.csv',
                 chunksize = 10000L,
                 todo = filter2)
                 }
}
