% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ora.R
\name{read_ora}
\alias{read_ora}
\title{Connection to the Oracle database}
\usage{
read_ora(
  dbname = "EMERALD",
  username = Sys.getenv("USERNAME"),
  password = rstudioapi::askForPassword(prompt = "Please enter your Oracle password...")
)
}
\arguments{
\item{dbname}{String with the name of the database. Defaults to "EMERALD".}

\item{username}{String with the name of the user. Defaults to current Microsoft User.}

\item{password}{String with the password of the user. Defaults to RStudio popup message.}
}
\value{
\code{oracon} which is a ROracle connection object.
}
\description{
Connection to the Oracle database
}
\examples{
\dontrun{oracon = read_ora()}
}
