% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_n_lines.R
\name{scan_n_lines}
\alias{scan_n_lines}
\title{Fast number of rows in text file}
\usage{
scan_n_lines(file)
}
\arguments{
\item{file}{The name of the text file to be scanned}
}
\value{
Integer of row number
}
\description{
Fast number of rows in text file
}
\examples{
scan_n_lines("test.csv")
}
