% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_n_lines.R
\name{scan_n_lines}
\alias{scan_n_lines}
\title{Fast number of rows in text file}
\usage{
scan_n_lines(file)
}
\arguments{
\item{file}{The path and name of the text file to be scanned.}
}
\value{
Integer of row number
}
\description{
This function uses the windows cmd to count the number of rows in a text file (csv, txt).
Linux and Mac OS are not supported yet. The text file should be on the local computer, the function
will not work on files from remote servers.
}
\examples{
scan_n_lines("C:/Users/USERNAME/Desktop/test.csv")
}
